Identify Events by type strings
Events now have a type string instead of a selector. This allows objects
that aren't interested in the event to just ignore it instead of die
when trying to call a method that doesn't exist.

This means that if a Component wants to respond to an event it needs to
be caught and handled in the receiveEvent method on that Component.