Create MovementSystem
So far this system deals with moving an entity to a particular
target at a given speed. As of now it responds to a walkTo notification
containing an entity and a target. This replaces the movement logic on
the LeftPlayer behavior. Because this seems to be the only reason LeftPlayer
existed, LeftPlayer was also removed.

When an entity arrives at a target it posts an arrivedAtTarget
notification. Right now the pathfinding system is the only one that
cares but I can see it being useful in the future for other things.